* Added: Create empty `config.ini` for easier user usage
* Added: Post install notes
* Added: Script to install directly from repository
* Added: Install needed components automatically after a Venus OS upgrade
* Changed: Added QML to `restoregui.sh`
* Changed: Bash output
* Changed: Moved ble part to `installble.sh`
* Changed: Optimized installation scripts
* Changed: Fixed black lint errors

+++ etc/dbus-serialbattery/restoregui.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
-set -x
-#restore original backup 
-cp -f /opt/victronenergy/gui/qml/PageBattery.qml.backup /opt/victronenergy/gui/qml/PageBattery.qml
+
+# remove comment for easier troubleshooting
+#set -x
+
+# restore original backup
+if [ -f /opt/victronenergy/gui/qml/PageBattery.qml.backup ]; then
+    cp -f /opt/victronenergy/gui/qml/PageBattery.qml.backup /opt/victronenergy/gui/qml/PageBattery.qml
+    echo "PageBattery.qml was restored."
+fi
+# restore original backup
+if [ -f /opt/victronenergy/gui/qml/PageLynxIonIo.qml.backup ]; then
+    cp -f /opt/victronenergy/gui/qml/PageLynxIonIo.qml.backup /opt/victronenergy/gui/qml/PageLynxIonIo.qml
+    echo "PageLynxIonIo.qml was restored."
+fi
 
 #stop gui
 svc -d /service/gui
